I'm having a hard time with the sqlite. I'm trying to understand how this table is put together and how you access everything.
I know this is only 1 table and it is joined with other tables. 

This program simply lists all the column names and what value is inside that column for that user. 

So far, I know you have to make a cursor item then pass it a uri. In this case you pass it the ContactsContract.Contacts.CONTACT_URI and it places the cursor on the first Contact in the list. 

http://developer.android.com/resources/articles/contacts.html you can see a little more how this works here. 

From there i do what is stated above for each contact. In my case i have 3.